vscode setupClient function
