+++ vscode-lean4/src/extension.ts
@@ -30,6 +30,7 @@ import { displayWarning } from './utils/notifs'
 import { PathExtensionProvider } from './utils/pathExtensionProvider'
 import { findLeanProjectRoot } from './utils/projectInfo'
 import { VSCodeInfoWebviewFactory } from './infowebview'
+import { setupClient } from './vscodeleanclient'
 
 async function setLeanFeatureSetActive(isActive: boolean) {
     await commands.executeCommand('setContext', 'lean4.isLeanFeatureSetActive', isActive)
@@ -159,7 +160,7 @@ async function activateLean4Features(
         return undefined
     }
 
-    const clientProvider = new LeanClientProvider(installer, installer.getOutputChannel())
+    const clientProvider = new LeanClientProvider(installer, installer.getOutputChannel(), setupClient)
     context.subscriptions.push(clientProvider)
 
     const watchService = new LeanConfigWatchService()
